requirements.txt
Flask>=3.0,<4.0
tensorflow>=2.12,<3.0
Pillow>=9.5
numpy>=1.23
gunicorn>=21.2
gevent>=24.2
python-dotenv>=1.0